make_audio_gemini/requirements.txt
protobuf==6.31.1
streamlit==1.47.1
google-genai
